Remove ONESHELL requirement for all make recipes
This removes the one shell requirement to have all make recipe lines
passed to a single shell invocation. While this can improve performance
and is useful for sharing environment variables across recipe lines
without `;` and escaping the newline character, once it is set anywhere
in the Makefile, then all make recipe lines for each target will be
provided to a single invocation of the shell. This removal will avoid
any confusion and keep the Makefile contents as consistent as possible
to the originally generated Makefile from operator-sdk.