make import-manifests using newer kustomize v3.8.6
A newer version of kustomize (v3.8.6) now generates a slightly different
order for certain fields in deployments and validating webhook
configurations, as well as different filenames for the core API group
resources. So after importing the manifests using this newer version of
kustomize, the existing core API group manifests are renamed to match
the new naming scheme.