Use operator version from Makefile
When invoking make bundle, the operator version passed in needs to match
the version in the CSV bundle. Therefore, it's best not to rely on a
hard-coded version as it will need to be updated each time a new release
of the operator is made because the CSV version will have been updated.
Instead of hard-coding the version, this change will use the version
that is stored in the Makefile, which is the canonical source for the
operator version and so will be kept updated for every release.