Set failurePolicy to Ignore to avoid e2e flakes
Setting failurePolicy to Fail causes issues when the Gatekeeper resource
is deleted and the Kubernetes garbage collector cleans up the resources
owned by the Gatekeeper resource in a nondeterministic fashion. This is
because the webhook Deployment may get cleaned up prior to the
ValidatingWebhookConfiguration and when the failurePolicy is set to
Fail, the webhook Deployment will be unreachable and will ultimately
block the successful deletion of the Gatekeeper resource and eventually
timeout.