Verify imported manifests using cloned gatekeeper
This adds steps to checkout the Gatekeeper repo alongside the Gatekeeper
Operator repo and use the Gatekeeper repo path as the path argument to
run kustomize build instead of using a go-getter URL with a git ref.
This also adds some workaround steps to get the manifests in proper
shape for building.