Reorganize e2e test files
This change moves the e2e test files under a new 'e2e' directory and
also removes the '_test' suffix from e2e test files that are not entry
points of the 'go test' command. That is, only one e2e file needs the
'_test' suffix as the entry point so that 'go test' can find and execute
it.