Do not skip creating the Namespace asset
When deploying Gatekeeper, we no longer skip updating the namespace
resource in Kubernetes (the namespace is already created), so that we
can apply the Gatekeeper namespace labels e.g.
'admission.gatekeeper.sh/ignore: no-self-managing'. We also avoid adding
an owner reference to the namespace resource so that deleting the
Gatekeeper CR does not delete the namespace and everything within it.