Run actions gatekeeper e2e tests directly
This change now checks out the Gatekeeper repository and runs the bats
e2e tests directly from the Gatekeeper repository instead of maintaining
a duplicate version of the bats tests in the operator repository. This
simplifies maintenance, and if it ever breaks, we can either revert back
to committing the bats tests in the operator repo, or fixing it directly
in the Gatekeeper version release branch.