Consistently use v in VERSION
Previously, "v" was in the VERSION variable. The operator-sdk scaffold
doesn't have that in its VERSION variable. However, more instances in
the Makefile need the "v", and many other *_VERSION vars also have a
"v", so this will make things more consistent.

The one case we don't want "v" is during the `generate bundle` step,
but it is easy to strip it from the variable there.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>