Fix RBAC  names and labels
As noted in the review, the commmonLabels and namePrefix should have
been kept in the config/rbac/base kustomization, since those are used
outside of the config/default kustomization.

Additionally, creating the bundle (in the next commit) made it obvious
that the service account name in the CSV needed to be updated as well.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>