Move Go setup earlier in workflows
The docker-build make target has other dependencies that use Go. Since
the specific version of Go was not set up until later in the Github
Action, I think the version it was using was whatever happened to
already be on that ubuntu VM. That could cause problems, possibly
including the recent image build failures.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>