Fix the missing role file when deploying on OCP
Previous versions of the Gatekeeper operator had a custom role
definition for OpenShift but as part of the upgrade to v3.11, it was
consolidated to a single role. It is benign to have an OpenShift
specific permission on other clusters since that API group does not
exist on other clusters. This keeps maintenance simpler.

Relates:
https://issues.redhat.com/browse/ACM-5295

Signed-off-by: mprahl <mprahl@users.noreply.github.com>